flask_requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
